Extract client method for packages, versions and deployment plan in PackageClient class
