+++ tests/clients/BarracksClient.spec.js
@@ -17,7 +17,6 @@ describe('Barracks', () => {
   let mockedCreateReadStream = undefined;
   let mockedBasename = undefined;
   const token = 'i8uhkj.token.65ryft';
-  const programWithValidOptions = {};
 
   beforeEach(() => {
     barracks = new Barracks();
@@ -66,6 +65,21 @@ describe('Barracks', () => {
       expect(barracksClient).to.have.property('deleteFilter').and.to.be.a('function');
     });
 
+    it('should initialize methods from packageClient when constructor called', () => {
+      // When
+      const barracksClient = new Barracks();
+
+      // Then
+      expect(barracksClient).to.have.property('createComponent').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getPackage').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getComponents').and.to.be.a('function');
+      expect(barracksClient).to.have.property('createVersion').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getVersion').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getComponentVersions').and.to.be.a('function');
+      expect(barracksClient).to.have.property('publishDeploymentPlan').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getDeploymentPlan').and.to.be.a('function');
+    });
+
     it('should initialize methods from segmentClient when constructor called', () => {
       // When
       const barracksClient = new Barracks();
@@ -103,266 +117,6 @@ describe('Barracks', () => {
       expect(barracksClient).to.have.property('publishUpdate').and.to.be.a('function');
       expect(barracksClient).to.have.property('archiveUpdate').and.to.be.a('function');
       expect(barracksClient).to.have.property('scheduleUpdate').and.to.be.a('function');
-    });
-  });
-
-  describe('#createComponent()', () => {
-
-    const componentRef = 'my.component.yo';
-    const componentName = 'A cool component';
-    const componentDescription = 'A very cool component';
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const component = { ref: componentRef, name: componentName, description: componentDescription };
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.createComponent(token, component).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'createComponent',
-          {
-            headers: { 'x-auth-token': token },
-            body: component
-          }
-        );
-        done();
-      });
-    });
-
-    it('should return the component created', done => {
-      // Given
-      const component = { ref: componentRef, name: componentName, description: componentDescription };
-      const response = { body: Object.assign({}, component, { id: 'theNewId' }) };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.createComponent(token, component).then(result => {
-        expect(result).to.be.equals(response.body);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'createComponent',
-          {
-            headers: { 'x-auth-token': token },
-            body: component
-          }
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#getComponent()', () => {
-
-    it('should return a stream object and deleguate to the client', done => {
-      // Given
-      const options = {
-        headers: { 'x-auth-token': token }
-      };
-
-      barracks.client.retrieveAllPages = sinon.spy();
-
-      // When / Then
-      barracks.getComponents(token).then(result => {
-        expect(result).to.be.instanceOf(PageableStream);
-        expect(barracks.client.retrieveAllPages).to.have.been.calledOnce;
-        expect(barracks.client.retrieveAllPages).to.have.been.calledWithExactly(
-          new PageableStream(),
-          'getComponents',
-          options,
-          'components'
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#createVersion()', () => {
-
-    const filename = 'file.txt';
-    const filePath = 'path/to/file.txt';
-    const version = {
-      id: '2.0',
-      name: 'version 2',
-      component: 'ref.package',
-      file: filePath,
-      description: 'description',
-      metadata: JSON.stringify({ data: 'value' })
-    };
-
-    let spyCreateReadStream;
-    const proxyCreateReadStream = (path) => {
-      spyCreateReadStream(path);
-      return filename;
-    };
-
-    let spyBasename;
-    const proxyBasename = (path) => {
-      spyBasename(path);
-      return filename;
-    };
-
-    const ProxifiedBarracks = proxyquire(barracksClientPath, {
-      fs: {
-        createReadStream: (path) => {
-          return proxyCreateReadStream(path);
-        }
-      },
-      path: {
-        basename: (path) => {
-          return proxyBasename(path);
-        }
-      }
-    });
-
-    const options = {
-      headers: {
-        'x-auth-token': token
-      },
-      formData: {
-        version: {
-          value: JSON.stringify({
-            id: version.id,
-            name: version.name,
-            description: version.description,
-            metadata: version.metadata
-          }),
-          options: {
-            contentType: 'application/json'
-          }
-        },
-        file: {
-          value: filename,
-          options: {
-            filename: filename,
-            contentType: 'application/octet-stream'
-          }
-        }
-      },
-      pathVariables: {
-        componentRef: version.component
-      }
-    };
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const proxyBarracks = new ProxifiedBarracks();
-      const error = { message: 'request failed' };
-      spyCreateReadStream = sinon.spy();
-      spyBasename = sinon.spy();
-      proxyBarracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      proxyBarracks.createVersion(token, version).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(proxyBarracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(proxyBarracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'createVersion',
-          options
-        );
-        done();
-      });
-    });
-
-    it('should return the serveur response body when request is successful', done => {
-      // Given
-      const proxyBarracks = new ProxifiedBarracks();
-      const response = { body: 'youpi created' };
-      spyCreateReadStream = sinon.spy();
-      spyBasename = sinon.spy();
-
-      proxyBarracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      proxyBarracks.createVersion(token, version).then(result => {
-        expect(result).to.be.equals(response.body);
-        expect(proxyBarracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(proxyBarracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'createVersion',
-          options
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#publishDeploymentPlan()', () => {
-
-    const packageRef = 'ze.ref';
-    const validPlan = {
-      package: packageRef,
-      data: {
-        some: 'value',
-        someOther: 'value'
-      }
-    };
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.publishDeploymentPlan(token, validPlan).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'publishDeploymentPlan',
-          {
-            headers: {
-              'x-auth-token': token
-            },
-            pathVariables: {
-              componentRef: packageRef
-            },
-            body: validPlan
-          }
-        );
-        done();
-      });
-    });
-
-    it('should return the plan created', done => {
-      // Given
-      const plan = Object.assign({}, validPlan, { id: 'kjcxse456tyhjkloiuytrfd' });
-      const response = { body: plan };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.publishDeploymentPlan(token, plan).then(result => {
-        expect(result).to.be.equals(response.body);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'publishDeploymentPlan',
-          {
-            headers: {
-              'x-auth-token': token
-            },
-            pathVariables: {
-              componentRef: packageRef
-            },
-            body: plan
-          }
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
     });
   });
 
@@ -528,97 +282,6 @@ describe('Barracks', () => {
     });
   });
 
-  describe('#getComponentVersions()', () => {
-
-    it('should return a stream object and deleguate to the client', done => {
-      // Given
-      const componentRef = 'my.component.ref';
-      const options = {
-        headers: {
-          'x-auth-token': token
-        },
-        pathVariables: { componentRef }
-      };
-      barracks.client.retrieveAllPages = sinon.spy();
-
-      // When / Then
-      barracks.getComponentVersions(token, componentRef).then(result => {
-        expect(result).to.be.instanceOf(PageableStream);
-        expect(barracks.client.retrieveAllPages).to.have.been.calledOnce;
-        expect(barracks.client.retrieveAllPages).to.have.been.calledWithExactly(
-          new PageableStream(),
-          'getComponentVersions',
-          options,
-          'versions'
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#getVersion()', () => {
-
-    const packageRef = 'io.barracks.package';
-    const versionId = '3.29.5';
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.getVersion(token, packageRef, versionId).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'getVersion',
-          {
-            headers: {
-              'x-auth-token': token
-            },
-            pathVariables: {
-              componentRef: packageRef,
-              versionId
-            }
-          }
-        );
-        done();
-      });
-    });
-
-    it('should return a token when request succeed', done => {
-      // Given
-      const version = { component: packageRef, id: versionId };
-      const response = { body: version };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getVersion(token, packageRef, versionId).then(result => {
-        expect(result).to.deep.equals(version);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly(
-          'getVersion',
-          {
-            headers: {
-              'x-auth-token': token
-            },
-            pathVariables: {
-              componentRef: packageRef,
-              versionId
-            }
-          }
-        );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
   describe('#checkUpdateAndDownload()', () => {
 
     const baseUrl = 'base/url';
@@ -734,112 +397,6 @@ describe('Barracks', () => {
         );
         expect(downloadSpy).to.have.been.calledOnce;
         expect(downloadSpy).to.have.been.calledWithExactly();
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#getDeploymentPlan()', () => {
-    const packageRef = 'ze.ref';
-    const validDeploymentPlan = {
-      package: packageRef,
-      data: {
-        some: 'value',
-        someOther: 'value'
-      }
-    };
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.getDeploymentPlan(token, validDeploymentPlan.package).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getDeploymentPlan', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            componentRef: validDeploymentPlan.package
-          }
-        });
-        done();
-      });
-    });
-
-    it('should return a deployment plan when request succeed', done => {
-      // Given
-      const response = { body: validDeploymentPlan };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getDeploymentPlan(token, validDeploymentPlan.package).then(result => {
-        expect(result).to.be.equals(validDeploymentPlan);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getDeploymentPlan', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            componentRef: validDeploymentPlan.package
-          }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#getPackage()', () => {
-    const packageRef = 'ze.ref';
-    const validPackage = {
-      id: "id",
-      user: "userId",
-      reference: packageRef,
-      description: "description",
-      name: "packageName"
-    };
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.getPackage(token, validPackage.reference).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getPackage', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            componentRef: validPackage.reference
-          }
-        });
-        done();
-      });
-    });
-
-    it('should return a package when request succeeds', done => {
-      // Given
-      const response = { body: validPackage };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getPackage(token, validPackage.reference).then(result => {
-        expect(result).to.be.equals(validPackage);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getPackage', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            componentRef: validPackage.reference
-          }
-        });
         done();
       }).catch(err => {
         done(err);
